Add ability to watch fs for endpoint updates (#9)
This change adds the option to specify an endpoints directory
to be used for deriving the discovery state. As pointed out in
the comments there is the `endpoints_watch_directory` which
when set causes the mock to read endpoints state from the
files system instead of using a static config.