Add mock identity service (#10)
## Motivation

The mock destination server has been particuarly useful in benchmarking changes
to `linkerd2-proxy`. This is because a single proxy can run without needing the
`linkerd2` control plane running; it sends all destination requests to the
server and uses the responses to send requests to the correct destinations.

A current limitation is that identity must be disabled on the proxy as seen
[here](https://github.com/olix0r/l2-proxy-harness/blob/main/inc.harness.sh#L106). This is because similar to the destination service, the identity service is 
also run as part of the control plane; without that there is way to get identity
for a specific service.

## Solution

Run an identity service that returns pre-signed certificates for specific
identity names.

A new `LINKERD2_MOCK_DST_IDENTITIES_DIR` environment variable can be set which
points to a directory that contains subdirectories--each corresponding to an
identity that should be served by the identity service.

When the identity service is created, it looks at each subdirectory which should
contain at least a `crt.pem`. (It's likely that it will also contain a `csr.der`
and `key.p8` which will be used by the proxy's `ENV_IDENTITY_DIR`.)

Taking the subdirectory name as the identity name, it will serve the
certificates in `crt.pem` for that directory name.

For example, for the `foo-ns1-ca1` identity, I have directory with the
following:

```bash
:; ls /path/to/identities/
ca-config.json ca1-key.pem ca1.pem foo-ns1-ca1/

:; ls /path/to/identities/foo-ns1-ca1/
crt.pem  csr.der  key.p8
```

When a request is made for the `foo-ns1-ca1` identity, the certifcates in
`crt.pem` are returned.

## Generating files

These all contain valid values and were generated similar to the proxy's test
harness `gen-certs.sh` script.

I lean towards including this script as part of the `l2-proxy-harness` project
where these certificates, CSRs, and keys will actually be used.

## Testing

This was tested with a gRPC client that sends a `CertifyRequest` for the
`foo-ns1-ca1` identity to the identity server running locally on `:8080`.

Gist [here](https://gist.github.com/kleimkuhler/e093aca197f89a91a1016b40ef6e7210)

Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>